Renaming indexer to prototype-indexer
