+++ contracts/sdk/prototype-indexer/indexer/log/utils.ts
