+++ contracts/sdk/prototype-indexer/indexer/utils.ts
