+++ contracts/sdk/prototype-indexer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "indexer",
+  "name": "prototype-indexer",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
