+++ contracts/sdk/prototype-indexer/smokeTest.ts
